requirements.txt
coincurve
pycryptodome
base58
fastecdsa